requirements.txt
requests
jira
pyyaml
